chore(AwsVpcPeering): refactoring to improve logging
